removed extra code, files
