--- TestShield/UnitTest1.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace TestShield
-{
-    [TestClass]
-    public class UnitTest1
-    {
-        [TestMethod]
-        public void TestMethod1()
-        {
-            //var stub = MockRepository.
-        }
-    }
-}
